Install protobuf-python from cached egg

+++ travis/install-protobuf.sh
@@ -14,7 +14,11 @@ fi
 
 if ! python -c "import google.protobuf.text_format"; then
     cd $HOME/protobuf/protobuf-2.6.1/python;
-    python setup.py install --cpp_implementation --user;
+    if [ -f "dist/protobuf-2.6.1-py2.7-linux-x86_64.egg" ]; then
+        python -m easy_install --user dist/protobuf-2.6.1-py2.7-linux-x86_64.egg
+    else
+        python setup.py install --cpp_implementation --user;
+    fi
     cd $HOME;
 else
     echo 'Using cached protobuf-python install.';
